backend/requirements.txt
anthropic>=0.34.0
langgraph>=0.2.0
langchain-anthropic>=0.2.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
redis>=5.1.0
qdrant-client>=1.11.0
openai>=1.50.0
tavily-python>=0.5.0
PyGithub>=2.4.0
APScheduler>=3.10.0
aiosqlite>=0.20.0
httpx>=0.27.0
python-dotenv>=1.0.0
sse-starlette>=2.0.0
python-telegram-bot>=21.0
mcp>=1.0.0
